requirements.txt
Django==1.4
--upgrade setuptools
numpy
nltk==2.0.4
distribute==0.6.21
Pillow
ply==3.4
pyaml
MySQL-python
django_nose
html5lib
beautifulsoup4
pisa==3.0.33
reportlab==2.7
# For sent_tokonizer manualy install punkt
South==1.0.2
